+++ recipes/ssh.rb
@@ -1,15 +1,3 @@
-node.default['openssh']['server']['password_authentication'] = 'no'
-node.default['openssh']['server']['challenge_response_authentication'] = 'no'
-node.default['openssh']['server']['pubkey_authentication'] = 'yes'
-node.default['openssh']['server']['x11_forwarding'] = 'no'
-
-# XXX: Check https://cipherli.st/ for latest recommendations
-node.default['openssh']['server']['ciphers'] = 'chacha20-poly1305@openssh.com,aes256-gcm@openssh.com,aes128-gcm@openssh.com,aes256-ctr,aes192-ctr,aes128-ctr'
-node.default['openssh']['server']['kex_algorithms'] = 'curve25519-sha256@libssh.org,diffie-hellman-group-exchange-sha256'
-node.default['openssh']['server']['m_a_cs'] = 'hmac-sha2-512-etm@openssh.com,hmac-sha2-256-etm@openssh.com,umac-128-etm@openssh.com,hmac-sha2-512,hmac-sha2-256,umac-128@openssh.com'
-
-include_recipe 'openssh::default'
-
 directory '/root/.ssh' do
   owner 'root'
   group 'root'
@@ -18,10 +6,31 @@ end
 
 ssh_authorized_keys = Array(node.fetch('sanity').fetch('ssh').fetch('authorized_keys'))
 
-file 'root/.ssh/authorized_keys' do
+file '/root/.ssh/authorized_keys' do
   owner 'root'
   group 'root'
   mode '0400'
   content ssh_authorized_keys.join("\n")
   not_if { ssh_authorized_keys.empty? }
 end
+
+node.override['ssh-hardening']['ssh']['server']['listen_to'] = \
+  if node['sanity']['ipv4_reachable']
+    node['ipaddress']
+  else
+    node['sanity']['ipv6'] ? '::' : '0.0.0.0'
+  end
+
+node.default['ssh-hardening']['network']['ipv6']['enable'] = node['sanity']['ipv6']
+
+%w(client server).each do |component|
+  # recommended crypto from https://github.com/dev-sec/ssh-baseline
+  node.default['ssh-hardening']['ssh'][component]['cipher'] = 'aes256-ctr,aes192-ctr,aes128-ctr'
+  node.default['ssh-hardening']['ssh'][component]['kex'] = 'curve25519-sha256,diffie-hellman-group-exchange-sha256'
+  node.default['ssh-hardening']['ssh'][component]['mac'] = 'hmac-sha2-512,hmac-sha2-256'
+end
+
+node.default['ssh-hardening']['ssh']['server']['allow_root_with_key'] = true
+node.default['ssh-hardening']['ssh']['use_privilege_separation'] = 'sandbox'
+
+include_recipe 'ssh-hardening::default'