+++ test/integration/default/inspec/monit_spec.rb
@@ -7,6 +7,7 @@ control 'monit-1' do
   end
 
   describe command('monit status') do
+    its('stderr') { should be_empty }
     its('stdout') { should match /^System '.+'$/ }
     its('stdout') { should match /^Filesystem '.+'$/ }
   end